// fe/types/employee.ts
export interface Employee {
  id: string
  avatar: string
  firstName: string
  lastName: string
  fullName: string
  jobTitle: string
  salary: number
  department: string
  location: string
}

// export enum JobTitle {
//   CEO = 'CEO',
//   CTO = 'CTO',
//   CFO = 'CFO',
//   Agent = 'Agent',
//   Manager = 'Manager',
//   SystemAdmin = 'System Admin',
//   SoftwareEngineer = 'Software Engineer',
//   HRAssociate = 'HR Associate',
//   HRSupervisor = 'HR Supervisor',
//   MarketingAssistant = 'Marketing Assistant',
//   MarketingLead = 'Marketing Lead',
// }

// export enum Department {
//   Board = 'Board',
//   Talent = 'Talent',
//   IT = 'IT',
//   HR = 'HR',
//   Marketing = 'Marketing',
// }

// export enum Location {
//   LosAngeles = 'Los Angeles',
//   NewYork = 'New York',
//   Miami = 'Miami',
// }